extra: add upstart-uevent-bridge.
This new bridge is designed to provide a replacement for udev bridge in
case of the system does not support for libudev. Some embedded systems
prefer the use of busybox mdev applet that is much more light than udev, or
even populate the device fs by the kernel itself).

The uevent brigde listens for the uevent kobjects sent via the netlink
socket. It emits the same upstart events that the udev bridge does:
 - *-device-added,
 - *-device-removed and
 - *-device-changed.

Signed-off-by: Gaël PORTAY <gael.portay@gmail.com>